Refactor solve()
This is in an attempt to make it clearer how the program attempts to execute a solve.  First a function will be called that attempts to remove candidates, then trySolve() will be called.  trySolve() tries to solve all inputs with only a single candidates, and returns true when if a solution is found for at least one input.  This in turn sets the bool solveFoundThisLoop to true.